Refactor into separate macros.
